requirements.txt
feedparser>=6.0.11
PyYAML>=6.0
python-dateutil>=2.8.0
requests>=2.28.0
trafilatura>=2.0.0
readability-lxml>=0.8.4.1
lxml>=5.0.0
beautifulsoup4>=4.12.0
